# ENH: sparse: speedup LIL matrix slicing