--- a/scipy/sparse/lil.py
+++ b/scipy/sparse/lil.py
@@ -7,11 +7,10 @@ __docformat__ = "restructuredtext en"
 
 __all__ = ['lil_matrix','isspmatrix_lil']
 
-from itertools import izip
 from bisect import bisect_left, bisect_right
 
 import numpy as np
-from scipy.lib.six import xrange
+from scipy.lib.six import xrange, zip as izip
 
 from .base import spmatrix, isspmatrix
 from .sputils import getdtype, isshape, issequence, isscalarlike, ismatrix, \