@rgommers 2015-01-01
This needs only a rebase and a very minor tweak it looks like. Time to get it in?