@perimosocordiae 2016-02-01
Rebased in #5789, closing.