@larsmans 2016-02-01
Reopening so that the status of this will be "merged" when I merge #5789.